address code comments, move new code to new file
